Move `AssignOps` into `Operators` lint pass
